requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.1=py39h2bbff1b_1
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=3.5.0=py39haa95532_0
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
boto3=1.21.32=pyhd3eb1b0_0
botocore=1.24.32=pyhd3eb1b0_0
bpemb=0.3.3=pypi_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py39h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2021.10.8=h5b45459_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py39hcbf5309_2
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.0=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
conllu=4.4.2=pypi_0
cryptography=36.0.0=py39h21b164f_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.28=py39hd77b12b_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecated=1.2.13=pypi_0
entrypoints=0.3=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.6.0=pypi_0
flair=0.11.1=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
frozenlist=1.2.0=py39h2bbff1b_0
ftfy=6.1.1=pypi_0
future=0.18.2=py39haa95532_1
gdown=3.12.2=pypi_0
gensim=4.1.2=py39hd77b12b_0
google-api-core=2.2.2=pyhd3eb1b0_0
google-auth=2.6.0=pyhd3eb1b0_0
google-cloud-core=2.2.2=pyhd3eb1b0_0
google-cloud-storage=2.1.0=pyh6c4a22f_0
google-crc32c=1.1.2=py39h2bbff1b_0
google-resumable-media=1.3.1=pyhd3eb1b0_1
googleapis-common-protos=1.53.0=py39h2eaa2aa_0
grpcio=1.42.0=py39hc60d5dd_0
hdbscan=0.8.28=py39h5d4886f_1
huggingface-hub=0.5.1=pypi_0
hyperopt=0.2.7=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=3.10.1=pypi_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.9.1=py39haa95532_0
ipython=8.2.0=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
janome=0.4.2=pypi_0
jedi=0.18.1=py39haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.2=py39haa95532_0
jupyter_server=1.13.5=pyhd3eb1b0_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.10.3=pyhd3eb1b0_1
kiwisolver=1.3.2=py39hd77b12b_0
konoha=4.6.5=pypi_0
langdetect=1.0.9=pypi_0
libcrc32c=1.1.1=ha925a31_2
libffi=3.4.2=h604cdb4_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.19.1=h23ce68f_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
llvmlite=0.38.0=py39h23ce68f_0
lxml=4.8.0=pypi_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.0.1=py39h2bbff1b_0
matplotlib=3.5.1=py39haa95532_1
matplotlib-base=3.5.1=py39hd77b12b_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
more-itertools=8.8.0=pypi_0
mpld3=0.3=pypi_0
multidict=5.1.0=py39h2bbff1b_2
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py39haa95532_1
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.1.0=py39haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.8=pypi_0
ninja=1.10.2=py39h559b2a2_3
nltk=3.7=pypi_0
notebook=6.4.8=py39haa95532_0
numba=0.55.1=py39hb8cd55e_0
numpy=1.21.5=py39ha4e8547_0
numpy-base=1.21.5=py39hc2deb75_0
openssl=1.1.1n=h8ffe710_0
overrides=3.1.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39h2e25243_1
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.7.0=py_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
plotly=5.6.0=pyhd3eb1b0_0
pptree=3.1=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.1=py39hd77b12b_0
pure_eval=0.2.2=pyhd3eb1b0_0
py4j=0.10.9.5=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pynndescent=0.5.6=pyh6c4a22f_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=py39haa95532_0
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pywin32=302=py39h827c3e9_1
pywinpty=2.0.2=py39h5da7b33_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py39hd77b12b_2
qt=5.9.7=vc14h73c81de_0
regex=2022.3.15=py39h2bbff1b_0
requests=2.27.1=pyhd3eb1b0_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.5.0=pyhd3eb1b0_0
sacremoses=0.0.50=pypi_0
scikit-learn=1.0.2=py39hf11a4ad_1
scipy=1.7.3=py39h0a974cb_0
segtok=1.5.11=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentence-transformers=2.2.0=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=62.1.0=py39hcbf5309_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.1.0=pyhd3eb1b0_0
sniffio=1.2.0=py39haa95532_1
sqlite=3.38.0=h2bbff1b_0
sqlitedict=2.0.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
tabulate=0.8.9=pypi_0
tbb=2021.5.0=h2d74725_1
tenacity=8.0.1=py39haa95532_0
terminado=0.13.1=py39haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torch=1.11.0=pypi_0
torchvision=0.12.0=pypi_0
tornado=6.1=py39h2bbff1b_0
tqdm=4.64.0=py39haa95532_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.18.0=pypi_0
typed-ast=1.4.3=py39h2bbff1b_1
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
umap-learn=0.5.3=py39hcbf5309_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
websocket-client=0.58.0=py39haa95532_4
wheel=0.37.1=pyhd3eb1b0_0
wikipedia-api=0.5.4=pypi_0
win_inet_pton=1.1.0=py39haa95532_0
winpty=0.4.3=4
xz=5.2.5=h62dcd97_0
yarl=1.6.3=py39h2bbff1b_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h8cc25b3_4
zstd=1.4.9=h19a0ad4_0
